Divide `state.hpp`
