+++ include/mpc/control/state.hpp
@@ -1,0 +1,5 @@
+/// @file state.hpp
+#pragma once
+#include <mpc/control/state/class.hpp>
+#include <mpc/control/state/state.hpp>
+#include <mpc/control/state/stateT.hpp>
