+++ include/mpc/control/state/stateT.hpp
@@ -1,14 +1,16 @@
-/// @file state.hpp
+/// @file stateT.hpp
 #pragma once
 #include <functional> // std::invoke
 #include <mpc/control/alternative.hpp>
 #include <mpc/control/monad.hpp>
+#include <mpc/control/state/class.hpp>
 #include <mpc/control/trans/class.hpp>
 #include <mpc/data/functor/identity.hpp>
 #include <mpc/prelude/compose.hpp>
 #include <mpc/prelude/fst.hpp>
-#include <mpc/prelude/unit.hpp>
 #include <mpc/type_traits.hpp>
+
+// clang-format off
 
 namespace mpc {
   // StateT
@@ -78,8 +80,6 @@ namespace mpc {
   // [x] applicative
   // [x] alternative
   // [x] monad_trans
-
-  // clang-format off
 
   /// instance (Monad m) => Monad (StateT s m) where
   ///     m >>= k  = StateT $ \ s -> do
@@ -322,23 +322,33 @@ namespace mpc {
 
     static constexpr lift_op lift{};
   };
-  // clang-format on
+
+  /// instance MonadState (StateT s) where
+  ///     state f = StateT (return . f)
+  template <copy_constructible_object Fn, class S>
+  struct monad_state_traits<StateT<Fn, S>> {
+    struct state_op {
+      template <copy_constructible_object Fn2>
+      constexpr auto operator()(Fn2&& f) const noexcept(
+        noexcept(   make_StateT<S>(compose(mpc::returns<StateT_monad_t<StateT<Fn, S>>>, std::forward<Fn2>(f)))))
+        -> decltype(make_StateT<S>(compose(mpc::returns<StateT_monad_t<StateT<Fn, S>>>, std::forward<Fn2>(f)))) {
+        return      make_StateT<S>(compose(mpc::returns<StateT_monad_t<StateT<Fn, S>>>, std::forward<Fn2>(f)));
+      }
+    };
+
+    static constexpr state_op state{};
+    static constexpr auto get1 = states::get1<StateT<Fn, S>>;
+    static constexpr auto put = states::put<StateT<Fn, S>>;
+  };
 
   // Grobal methods:
   // [x] eval_StateT
   // [x] exec_StateT
   // [x] map_StateT
   // [x] with_StateT
-  // [x] state
-  // [x] get1
-  // [x] put
-  // [x] modify
-  // [ ] modify'
-  // [x] get2
 
-  // eval_StateT, exec_StateT, map_StateT, with_StateT, state
+  // eval_StateT, exec_StateT, map_StateT, with_StateT
   namespace detail {
-    // clang-format off
     struct eval_StateT_op {
       template <isStateT ST, class T>
       constexpr auto operator()(ST&& x, T&& t) const noexcept(
@@ -374,20 +384,6 @@ namespace mpc {
         return      make_StateT<StateT_state_t<ST>>(compose(run_StateT % std::forward<ST>(x), std::forward<Fn2>(f)));
       }
     };
-
-    template <class>
-    struct state_op;
-
-    template <copy_constructible_object Fn, class S>
-    struct state_op<StateT<Fn, S>> {
-      template <copy_constructible_object Fn2>
-      constexpr auto operator()(Fn2&& f) const noexcept(
-        noexcept(   make_StateT<S>(compose(mpc::returns<StateT_monad_t<StateT<Fn, S>>>, std::forward<Fn2>(f)))))
-        -> decltype(make_StateT<S>(compose(mpc::returns<StateT_monad_t<StateT<Fn, S>>>, std::forward<Fn2>(f)))) {
-        return      make_StateT<S>(compose(mpc::returns<StateT_monad_t<StateT<Fn, S>>>, std::forward<Fn2>(f)));
-      }
-    };
-    // clang-format on
   } // namespace detail
 
   inline namespace cpo {
@@ -398,184 +394,7 @@ namespace mpc {
     inline constexpr perfect_forwarded_t<detail::map_StateT_op> map_StateT{};
 
     inline constexpr perfect_forwarded_t<detail::with_StateT_op> with_StateT{};
-
-    template <isStateT ST>
-    inline constexpr perfect_forwarded_t<detail::state_op<std::remove_cvref_t<ST>>> state{};
-  } // namespace cpo
-
-  // get1, put, modify, get2
-  namespace detail {
-    // clang-format off
-    template <isStateT ST>
-    struct put_op {
-      struct closure {
-        template <class T, class A>
-        constexpr auto operator()(T&& t, A&&) const noexcept(
-          noexcept(   std::make_pair(unit, std::forward<T>(t))))
-          -> decltype(std::make_pair(unit, std::forward<T>(t))) {
-          return      std::make_pair(unit, std::forward<T>(t));
-        }
-      };
-
-      template <class T>
-      constexpr auto operator()(T&& t) const noexcept(
-        noexcept(   mpc::state<ST>(perfect_forwarded_t<closure>{}(std::forward<T>(t)))))
-        -> decltype(mpc::state<ST>(perfect_forwarded_t<closure>{}(std::forward<T>(t)))) {
-        return      mpc::state<ST>(perfect_forwarded_t<closure>{}(std::forward<T>(t)));
-      }
-    };
-
-    template <isStateT ST>
-    struct modify_op {
-      struct closure {
-        template <class Fn, class T>
-        constexpr auto operator()(Fn&& f, T&& t) const noexcept(
-          noexcept(   std::make_pair(unit, std::invoke(std::forward<Fn>(f), std::forward<T>(t)))))
-          -> decltype(std::make_pair(unit, std::invoke(std::forward<Fn>(f), std::forward<T>(t)))) {
-          return      std::make_pair(unit, std::invoke(std::forward<Fn>(f), std::forward<T>(t)));
-        }
-      };
-
-      template <class Fn>
-      constexpr auto operator()(Fn&& f) const noexcept(
-        noexcept(   mpc::state<ST>(perfect_forwarded_t<closure>{}(std::forward<Fn>(f)))))
-        -> decltype(mpc::state<ST>(perfect_forwarded_t<closure>{}(std::forward<Fn>(f)))) {
-        return      mpc::state<ST>(perfect_forwarded_t<closure>{}(std::forward<Fn>(f)));
-      }
-    };
-
-    template <isStateT ST>
-    struct get2_op {
-      struct closure {
-        template <class Fn, class T>
-        constexpr auto operator()(Fn&& f, const T& t) const noexcept(
-          noexcept(   std::make_pair(std::invoke(std::forward<Fn>(f), t), t)))
-          -> decltype(std::make_pair(std::invoke(std::forward<Fn>(f), t), t)) {
-          return      std::make_pair(std::invoke(std::forward<Fn>(f), t), t);
-        }
-      };
-
-      template <class Fn>
-      constexpr auto operator()(Fn&& f) const noexcept(
-        noexcept(   mpc::state<ST>(perfect_forwarded_t<closure>{}(std::forward<Fn>(f)))))
-        -> decltype(mpc::state<ST>(perfect_forwarded_t<closure>{}(std::forward<Fn>(f)))) {
-        return      mpc::state<ST>(perfect_forwarded_t<closure>{}(std::forward<Fn>(f)));
-      }
-    };
-    // clang-format on
-  } // namespace detail
-
-  inline namespace cpo {
-    template <isStateT ST>
-    inline constexpr auto get1 = mpc::state<ST>([](const auto& t) { return std::make_pair(t, t); });
-
-    template <isStateT ST>
-    inline constexpr perfect_forwarded_t<detail::put_op<std::remove_cvref_t<ST>>> put{};
-
-    template <isStateT ST>
-    inline constexpr perfect_forwarded_t<detail::modify_op<std::remove_cvref_t<ST>>> modify{};
-
-    template <isStateT ST>
-    inline constexpr perfect_forwarded_t<detail::get2_op<std::remove_cvref_t<ST>>> get2{};
-  } // namespace cpo
-
-  // State
-  // [x] State
-  // [x] isState
-  // [x] State_state_t
-  // [ ] State_monad_t
-  // [x] make_State
-  // [x] run_State
-
-  /// type State s = StateT s Identity
-  template <copy_constructible_object Fn, class S>
-  requires std::invocable<Fn, S> and isIdentity<std::invoke_result_t<Fn, S>>
-  using State = StateT<Fn, S>;
-
-  // isState
-  namespace detail {
-    template <class>
-    struct is_State : std::false_type {};
-
-    template <copy_constructible_object Fn, class S>
-    requires std::invocable<Fn, S> and isIdentity<std::invoke_result_t<Fn, S>>
-    struct is_State<StateT<Fn, S>> : std::true_type {
-    };
-  } // namespace detail
-
-  template <class T>
-  concept isState = detail::is_State<std::remove_cvref_t<T>>::value;
-
-  template <isState ST>
-  using State_state_t = typename std::remove_cvref_t<ST>::state_type;
-
-  // This seems to be unnecessary because we know `State_monad_t = Identity`
-  // template<isState ST>
-  // using State_monad_t = typename std::remove_cvref_t<ST>::monad_type;
-
-  // make_State, run_State
-  namespace detail {
-    struct run_State_op {
-      template <isState ST>
-      constexpr auto operator()(ST&& x) const noexcept
-        -> decltype(compose(run_Identity, run_StateT % std::forward<ST>(x))) {
-        return compose(run_Identity, run_StateT % std::forward<ST>(x));
-      }
-    };
-  } // namespace detail
-
-  namespace cpo {
-    template <class S>
-    inline constexpr auto make_State = make_StateT<S>;
-
-    inline constexpr perfect_forwarded_t<detail::run_State_op> run_State{};
-  } // namespace cpo
-
-  // Grobal methods:
-  // [x] eval_State
-  // [x] exec_State
-  // [x] map_State
-  // [x] with_State
-
-  // eval_State, exec_State, map_State, with_State
-  namespace detail {
-    // clang-format off
-    struct eval_State_op {
-      template <isState ST, class T>
-      constexpr auto operator()(ST&& x, T&& t) const noexcept(
-        noexcept(   fst(run_State % std::forward<ST>(x) % std::forward<T>(t))))
-        -> decltype(fst(run_State % std::forward<ST>(x) % std::forward<T>(t))) {
-        return      fst(run_State % std::forward<ST>(x) % std::forward<T>(t));
-      }
-    };
-
-    struct exec_State_op {
-      template <isState ST, class T>
-      constexpr auto operator()(ST&& x, T&& t) const noexcept(
-        noexcept(   snd(run_State % std::forward<ST>(x) % std::forward<T>(t))))
-        -> decltype(snd(run_State % std::forward<ST>(x) % std::forward<T>(t))) {
-        return      snd(run_State % std::forward<ST>(x) % std::forward<T>(t));
-      }
-    };
-
-    struct map_State_op {
-      template <class Fn2, isState ST>
-      constexpr auto operator()(Fn2&& f, ST&& x) const noexcept(
-        noexcept(   map_StateT % compose(make_Identity, compose(std::forward<Fn2>(f), run_Identity)) % std::forward<ST>(x)))
-        -> decltype(map_StateT % compose(make_Identity, compose(std::forward<Fn2>(f), run_Identity)) % std::forward<ST>(x)) {
-        return      map_StateT % compose(make_Identity, compose(std::forward<Fn2>(f), run_Identity)) % std::forward<ST>(x);
-      }
-    };
-    // clang-format on
-  } // namespace detail
-
-  inline namespace cpo {
-    inline constexpr perfect_forwarded_t<detail::eval_State_op> eval_State{};
-
-    inline constexpr perfect_forwarded_t<detail::exec_State_op> exec_State{};
-
-    inline constexpr perfect_forwarded_t<detail::map_State_op> map_State{};
-
-    inline constexpr auto with_State = with_StateT;
   } // namespace cpo
 } // namespace mpc
+
+// clang-format on
