+++ tests/control/control.cpp
@@ -4,4 +4,4 @@
 #include <mpc/data.hpp>
 #include "../stdfundamental.hpp"
 
-TEST_CASE("control", "[control]") {}
+// TEST_CASE("control", "[control]") {}
